Fix garbage held key state after using open file dialog
This removes some nice key sync stuff we were doing, but it doesn't
seem to be reliable due to sometimes getting junk out of GetKeyState()
due to some kind of Win32 quirk. Truly solving it seems like it would
require a lot of code, and it's easier to just clear the held keys
when the window lsoes focus.