Add deferred initialization for WaveOut audio
waveOutOpen() can block for 50ms or more, which causes a delay before
the emulator window appears if the ROM uses audio immediately. Instead,
we'll use a timer to defer audio initialization for 100ms after the first
paint.

Kinda nasty, but I couldn't think of a better way to handle this. The timer
works better than PostMessage because there can still be some jank with
the taskbar icon (and probably other things I didn't notice.)