Shuffle CPU opcode values to match the shuffling the uxnemu project did
This breaks compatibility with all existing Uxn ROMs. Sorry.

You will need to recompile all of your ROMs to work with the new opcode
assignments. Source code changes to .tal files are probably not needed.

The uxnemu project didn't state a specific reason as to why this was
done, but Uxn32 has to change to match the new opcode ordering,
otherwise Uxn32 will be incompatible going forward.

Sorry about that.