Add special-casing to display instr 0x00 as BRK, LITk as LIT
This commit changes DecodeUxnOpcode() (which should probably be called
InstructionToMnemonic) so that instruction 0x00 will be displayed as BRK
instead of LIT, and so that any LIT instructions (0x00 plus at least one
non-zero bit at positions 6, 7, or 8) are displayed without the 'k' modifier,
matching the behavior of Uxn assemblers.

This commit also changes EncodeUxnOpcode() so that the these rules work when
going the other direction, from text to instruction byte value.