Change 16-bit POKE/PEEK at 0xFFFF to operate on a padding byte
A while back, we had changed Uxn32 so that 16-bit operations with
destination address 0xFFFF would wrap the higher byte to 0x0.
This commit reverts that behavior, so that it matches how uxnsdl
works. But instead of writing and reading from unknown host memory,
we also add an additional byte of padding to the allocated Uxn RAM
size. Now, there's an extra secret byte of memory there.