Add handling of 0xF 'exit' flag in 'system' device
This is supposed to let Uxn programs 'quit'. In uxnsdl, this makes the
emulator window close immediately. This seems a bit weird for Uxn32, so I
made it open the debugger and display a message saying that the program
requested exit. I'm not sure if this is a good idea or not.