Add buffering for the virtual console device
This speeds up writing to the Uxn virtual console by keeping a separate buffer
of chars outside of the Win32 managed console window. Before, for simplicity,
we just appended every character one at a time to the console window. Now, we
use a timer to flush the char buffer to the console window after a short
period, or when the buffer becomes full.