Rename the Uint8, Sint8, and Uint16 types in core32.h
They are now named UxnU8, UxnI8, and UxnU16. This change is intended to make
it easier for other people to use the core Uxn implementation from Uxn32
without worrying about name collisions.

The consistent 'Uxn' prefix makes it easier to avoid name collisions. The
previous Uint8, Sin8, and Uint16 names are a bit common, and had a higher
chance of colliding in other people's existing projects.