Fix PC getting out of bounds during LDR and STR instructions
We were just using uint for it on the stack, but the manual masking
required to deal with the new call instructions was going to be too
annoying to be worth it. Just use U16 for the stack variable, so that
we get C's wrapping guarantee for unsigned, and then mask it in the two
places where arithmetic could push it out of Uxn RAM bounds.