Disable file dev's ability to use certain file extensions
If the Uxn program attempts to set the path to have the extension .exe, .dll,
or .com, it will now result in an error. This is a preventative measure
against .dll hijacking and modifying or creating .exe files. Files with these
extensions are still visible in the files listing, though.