Change 16-bit zero-page LDZ and STZ to wrap within zero-page.
This commit fixes compatibility with the newer Uxn specification.
Previously, LDZ and STZ in 16-bit mode could have the second byte read or
write beyond the zero-page address range. After changing the spec,
LDZ and STZ have the second byte wrap to 0x0000.

This commit adds a fair amount of noise to the uxn_core.c code. I had to
parameterize the address wrapping mask on POKE, PEEK16, and PEEK macros
so that they could be used in both 16-bit wrap mode (normal mode) and
8-bit wrap mode (for LDZ and STZ.)

I'm not sure if I like this solution, but let's go with it for now.
It won't be any slower, but it is harder to read the source code.