Change wst/rst depth fetch device ports, add setting depth
This is to match the new Uxn specification behavior.

This commit also ifdefs-out window title setting, since it occupied one
of the port numbers we needed, and has been moved to the expansion device
in the newer Uxn specification. We will need to modify our implementation
to work like that, in the future.