Update stasher (system memory device) to modern Uxn specification
This commit also removes the stasher fault code and fault handling,
because it's no longer possible for the stasher to fault.

Unfortunately, the new specification is more complicated than the old
one, and requires more code and code paths. It's also slower by
default, because it's specified in terms of two different directions of
byte-by-byte copying with wrapping. To try to avoid the stasher bank
switching from becoming much slower, there is a memcpy code path when
the two memory regions don't overlap and don't wrap.

I don't really like this code, but I'm not sure there is a good way to
implement this new specification, given the limitations Uxn32 needs to
conform to.